+++ src/ChatContext.tsx
@@ -1,18 +1,6 @@
 import { createContext, useContext, useEffect, useRef, useState } from "react";
-import {
-  ChatContextType,
-  ChatProviderType,
-  FileAttachment,
-  GroupChatEventType,
-  UnreadMessagesCount,
-} from "./types";
-import {
-  Chat,
-  DateOrTimestamp,
-  Dialogs,
-  Messages,
-  Users,
-} from "connectycube/dist/types/types";
+import { ChatContextType, ChatProviderType, GroupChatEventType, UnreadMessagesCount } from "./types";
+import { Chat, DateOrTimestamp, Dialogs, Messages, Users } from "connectycube/dist/types/types";
 import ConnectyCube from "connectycube";
 import useStateRef from "react-usestateref";
 import { formatDistanceToNow } from "date-fns";
@@ -30,28 +18,19 @@ export const useChat = (): ChatContextType => {
   return context;
 };
 
-export const ChatProvider = ({
-  children,
-}: ChatProviderType): React.ReactElement => {
-  const [isOnline, setIsOnline, _isOnlineRef] = useStateRef<boolean>(
-    navigator.onLine
-  );
-  const [currentUserId, setCurrentUserId, currentUserIdRef] = useStateRef<
-    number | undefined
-  >();
+export const ChatProvider = ({ children }: ChatProviderType): React.ReactElement => {
+  const [isOnline, setIsOnline, _isOnlineRef] = useStateRef<boolean>(navigator.onLine);
+  const [currentUserId, setCurrentUserId, currentUserIdRef] = useStateRef<number | undefined>();
   const [isConnected, setIsConnected] = useState(false);
   const [dialogs, setDialogs, dialogsRef] = useStateRef<Dialogs.Dialog[]>([]);
-  const [unreadMessagesCount, setUnreadMessagesCount] =
-    useState<UnreadMessagesCount>({ total: 0 });
+  const [unreadMessagesCount, setUnreadMessagesCount] = useState<UnreadMessagesCount>({ total: 0 });
   const [users, setUsers, usersRef] = useStateRef<{
     [userId: number]: Users.User;
   }>({});
   const [_onlineUsers, setOnlineUsers, onlineUsersRef] = useStateRef<
     Users.UsersResponse & Users.ListOnlineParams & { requested_at: number }
   >({ users: [], limit: 100, offset: 0, requested_at: 0 });
-  const [selectedDialog, setSelectedDialog] = useState<
-    Dialogs.Dialog | undefined
-  >();
+  const [selectedDialog, setSelectedDialog] = useState<Dialogs.Dialog | undefined>();
   const [messages, setMessages, messagesRef] = useStateRef<{
     [dialogId: string]: Messages.Message[];
   }>({});
@@ -61,10 +40,9 @@ export const ChatProvider = ({
   const [typingStatus, setTypingStatus] = useState<{
     [dialogId: string]: { [userId: string]: boolean };
   }>({});
-  const [activatedDialogs, setActivatedDialogs, _activatedDialogsRef] =
-    useStateRef<{
-      [dialogId: string]: boolean;
-    }>({});
+  const [activatedDialogs, setActivatedDialogs, _activatedDialogsRef] = useStateRef<{
+    [dialogId: string]: boolean;
+  }>({});
   const typingTimers = useRef<{ [key: string]: NodeJS.Timeout }>({});
   const onMessageRef = useRef<Chat.OnMessageListener | null>(null);
 
@@ -105,10 +83,7 @@ export const ChatProvider = ({
     return newDialog;
   };
 
-  const createGroupChat = async (
-    usersIds: number[],
-    chatName: string
-  ): Promise<Dialogs.Dialog> => {
+  const createGroupChat = async (usersIds: number[], chatName: string): Promise<Dialogs.Dialog> => {
     const params = {
       type: 2,
       name: chatName,
@@ -128,9 +103,7 @@ export const ChatProvider = ({
     return dialog;
   };
 
-  const getDialogs = async (
-    filters?: Dialogs.ListParams
-  ): Promise<Dialogs.Dialog[]> => {
+  const getDialogs = async (filters?: Dialogs.ListParams): Promise<Dialogs.Dialog[]> => {
     // fetch chats
     const result = await ConnectyCube.chat.dialog.list(filters);
 
@@ -144,28 +117,18 @@ export const ChatProvider = ({
         uniqueDialogsMap.set(dialog._id, dialog);
       });
       // Convert the map values back to an array and sort it
-      const uniqueDialogs = Array.from(uniqueDialogsMap.values()).sort(
-        (a, b) => {
-          const dateA =
-            _parseDate(a.last_message_date_sent) ||
-            _parseDate(a.created_at) ||
-            0;
-          const dateB =
-            _parseDate(b.last_message_date_sent) ||
-            _parseDate(b.created_at) ||
-            0;
+      const uniqueDialogs = Array.from(uniqueDialogsMap.values()).sort((a, b) => {
+        const dateA = _parseDate(a.last_message_date_sent) || _parseDate(a.created_at) || 0;
+        const dateB = _parseDate(b.last_message_date_sent) || _parseDate(b.created_at) || 0;
 
-          return dateB - dateA; // Sort in descending order (most recent first)
-        }
-      );
+        return dateB - dateA; // Sort in descending order (most recent first)
+      });
 
       return uniqueDialogs;
     });
 
     // store users
-    const usersIds = Array.from(
-      new Set(result.items.flatMap((dialog) => dialog.occupants_ids))
-    );
+    const usersIds = Array.from(new Set(result.items.flatMap((dialog) => dialog.occupants_ids)));
     _retrieveAndStoreUsers(usersIds);
 
     return result.items;
@@ -186,12 +149,11 @@ export const ChatProvider = ({
         return a._id.toString().localeCompare(b._id.toString()); // revers sort
       })
       .map((msg) => {
-        const attachmentsUrls = msg.attachments?.map((attachment) => {
-          const fileUrl = ConnectyCube.storage.privateUrl(attachment.uid);
-          return fileUrl;
-        });
-
-        return { ...msg, attachmentsUrls };
+        const attachments = msg.attachments?.map((attachment) => ({
+          ...attachment,
+          url: ConnectyCube.storage.privateUrl(attachment.uid),
+        }));
+        return { ...msg, attachments };
       });
     setMessages({ ...messages, [dialogId]: retrievedMessages });
 
@@ -287,16 +249,12 @@ export const ChatProvider = ({
     _retrieveAndStoreUsers(usersIds);
     //
     const updatedDialog = _findDialog(selectedDialog._id);
-    updatedDialog.occupants_ids = Array.from(
-      new Set([...updatedDialog.occupants_ids, ...usersIds])
-    );
+    updatedDialog.occupants_ids = Array.from(new Set([...updatedDialog.occupants_ids, ...usersIds]));
     setSelectedDialog({ ...updatedDialog });
     setDialogs([...dialogs]);
   };
 
-  const removeUsersFromGroupChat = async (
-    usersIds: number[]
-  ): Promise<void> => {
+  const removeUsersFromGroupChat = async (usersIds: number[]): Promise<void> => {
     if (!selectedDialog) {
       throw new Error("No dialog selected");
     }
@@ -323,9 +281,7 @@ export const ChatProvider = ({
 
     // update store
     const updatedDialog = _findDialog(selectedDialog._id);
-    updatedDialog.occupants_ids = updatedDialog.occupants_ids.filter(
-      (userId) => !usersIds.includes(userId)
-    );
+    updatedDialog.occupants_ids = updatedDialog.occupants_ids.filter((userId) => !usersIds.includes(userId));
     setSelectedDialog({ ...updatedDialog });
     setDialogs([...dialogs]);
   };
@@ -341,11 +297,7 @@ export const ChatProvider = ({
     selectedDialog.occupants_ids
       .filter((userId) => userId !== currentUserId)
       .forEach((userId) => {
-        _notifyUsers(
-          GroupChatEventType.REMOVED_FROM_DIALOG,
-          selectedDialog._id,
-          userId
-        );
+        _notifyUsers(GroupChatEventType.REMOVED_FROM_DIALOG, selectedDialog._id, userId);
       });
 
     setDialogs(dialogs.filter((dialog) => dialog._id !== selectedDialog._id));
@@ -362,75 +314,61 @@ export const ChatProvider = ({
     const messageId = _sendMessage(body, null, dialog, opponentId);
 
     // add message to store
-    _addMessageToStore(
-      messageId,
-      body,
-      dialog._id,
-      currentUserId as number,
-      opponentId
-    );
+    _addMessageToStore(messageId, body, dialog._id, currentUserId as number, opponentId);
   };
 
-  const sendMessageWithAttachment = async (
-    file: File,
-    dialog?: Dialogs.Dialog
-  ): Promise<void> => {
+  const sendMessageWithAttachment = async (files: File[], dialog?: Dialogs.Dialog): Promise<void> => {
     dialog ??= selectedDialog;
     if (!dialog) {
       throw "No dialog provided. You need to provide a dialog via function argument or select a dialog via 'selectDialog'.";
     }
 
     const opponentId = getDialogOpponentId(dialog);
-
     const tempId = Date.now() + "";
+    const attachments = files.map((file, index) => ({
+      uid: `local:${tempId}#${index}`, // just for temporary
+      type: file.type,
+      url: URL.createObjectURL(file),
+    }));
 
     // add message to store
-    const localUrl = URL.createObjectURL(file);
-    _addMessageToStore(
-      tempId,
-      "Attachment", //file.type,
-      dialog._id,
-      currentUserId as number,
-      opponentId,
-      localUrl,
-      true
-    );
+    _addMessageToStore(tempId, "Attachment", dialog._id, currentUserId as number, opponentId, attachments, true);
 
-    // upload file to cloud
-    const fileParams = {
-      name: file.name,
-      file: file,
-      type: file.type,
-      size: file.size,
-      public: false,
-    };
-    const result = await ConnectyCube.storage.createAndUpload(fileParams);
+    // upload files to cloud
+    const uploadFilesPromises = files.map((file) => {
+      const fileParams = {
+        file: file,
+        name: file.name,
+        type: file.type,
+        size: file.size,
+        public: false,
+      };
+      return ConnectyCube.storage.createAndUpload(fileParams);
+    });
+
+    const uploadedFilesResults = await Promise.all(uploadFilesPromises);
+    const uploadedAttachments = uploadedFilesResults.map(({ uid, type }) => ({
+      uid,
+      type,
+      url: ConnectyCube.storage.privateUrl(uid),
+    }));
 
     // send
-    const messageId = _sendMessage(
-      "Attachment",
-      { uid: result.uid as string, type: file.type },
-      dialog,
-      opponentId
-    );
-
-    const fileUrl = ConnectyCube.storage.privateUrl(result.uid);
+    const messageId = _sendMessage("Attachment", uploadedAttachments, dialog, opponentId);
 
     // update message in store (update it and file url)
-    const msg = messagesRef.current[dialog._id].find(
-      (msg) => msg._id === tempId
-    ) as Messages.Message;
+    const msg = messagesRef.current[dialog._id].find((msg) => msg._id === tempId) as Messages.Message;
     msg._id = messageId;
-    msg.attachmentsUrls = [fileUrl];
+    msg.attachments = attachments;
     msg.isLoading = false;
     setMessages({ ...messagesRef.current });
   };
 
   const _sendMessage = (
     body: string,
-    fileParams: FileAttachment | null,
+    attachments: Messages.Attachment[] | null,
     dialog: Dialogs.Dialog,
-    opponentId?: number
+    opponentId?: number,
   ): string => {
     // send message
     const messageParams: Chat.MessageParams = {
@@ -441,15 +379,10 @@ export const ChatProvider = ({
         dialog_id: dialog._id,
       },
     };
-    if (fileParams) {
-      messageParams.extension.attachments = [
-        { uid: fileParams.uid, type: fileParams.type },
-      ];
+    if (attachments) {
+      messageParams.extension.attachments = attachments;
     }
-    const messageId = ConnectyCube.chat.send(
-      dialog.type === 3 ? (opponentId as number) : dialog._id,
-      messageParams
-    );
+    const messageId = ConnectyCube.chat.send(dialog.type === 3 ? (opponentId as number) : dialog._id, messageParams);
 
     return messageId;
   };
@@ -460,8 +393,8 @@ export const ChatProvider = ({
     dialogId: string,
     senderId: number,
     recipientId?: number,
-    fileUrl?: string,
-    isLoading?: boolean
+    attachments?: Messages.Attachment[],
+    isLoading?: boolean,
   ) => {
     const ts = Math.round(new Date().getTime() / 1000);
 
@@ -472,15 +405,11 @@ export const ChatProvider = ({
     dialog.last_message_date_sent = ts;
     setDialogs(
       [...dialogsRef.current].sort((a, b) => {
-        const dateA =
-          _parseDate(a.last_message_date_sent) ||
-          (_parseDate(a.created_at) as number);
-        const dateB =
-          _parseDate(b.last_message_date_sent) ||
-          (_parseDate(b.created_at) as number);
+        const dateA = _parseDate(a.last_message_date_sent) || (_parseDate(a.created_at) as number);
+        const dateB = _parseDate(b.last_message_date_sent) || (_parseDate(b.created_at) as number);
 
         return dateB - dateA; // Sort in ascending order
-      })
+      }),
     );
 
     setMessages({
@@ -500,8 +429,7 @@ export const ChatProvider = ({
           read_ids: [senderId],
           delivered_ids: [senderId],
           views_count: 0,
-          attachments: [],
-          attachmentsUrls: fileUrl ? [fileUrl] : [],
+          attachments: attachments ? attachments : [],
           reactions: {} as any,
           isLoading,
         },
@@ -531,9 +459,7 @@ export const ChatProvider = ({
         message.read = 1;
         const dialog = _findDialog(dialogId);
         if (dialog) {
-          dialog.unread_messages_count >= 1
-            ? dialog.unread_messages_count--
-            : 0;
+          dialog.unread_messages_count >= 1 ? dialog.unread_messages_count-- : 0;
         }
 
         setDialogs([...dialogs]);
@@ -551,24 +477,16 @@ export const ChatProvider = ({
       };
       const result = await ConnectyCube.users.getV2(params);
 
-      const usersIdsMap = result.items.reduce<{ [key: number]: Users.User }>(
-        (map, user) => {
-          map[user.id] = user;
-          return map;
-        },
-        {}
-      );
+      const usersIdsMap = result.items.reduce<{ [key: number]: Users.User }>((map, user) => {
+        map[user.id] = user;
+        return map;
+      }, {});
 
       setUsers({ ...usersRef.current, ...usersIdsMap });
     }
   };
 
-  const _notifyUsers = (
-    command: string,
-    dialogId: string,
-    userId: number,
-    params: any = {}
-  ) => {
+  const _notifyUsers = (command: string, dialogId: string, userId: number, params: any = {}) => {
     const msg = {
       body: command,
       extension: {
@@ -587,35 +505,28 @@ export const ChatProvider = ({
       full_name: { start_with: term },
       limit: 100,
     });
-    users.push(
-      ...usersWithFullName.items.filter((user) => user.id !== currentUserId)
-    );
+    users.push(...usersWithFullName.items.filter((user) => user.id !== currentUserId));
 
     const usersWithLogin = await ConnectyCube.users.getV2({
       login: { start_with: term },
       limit: 100,
     });
-    users.push(
-      ...usersWithLogin.items.filter((user) => user.id !== currentUserId)
-    );
+    users.push(...usersWithLogin.items.filter((user) => user.id !== currentUserId));
 
     // remove duplicates and current user for search
     return users
-      .filter(
-        (user, ind) => ind === users.findIndex((elem) => elem.id === user.id)
-      )
+      .filter((user, ind) => ind === users.findIndex((elem) => elem.id === user.id))
       .filter((user) => user.id !== parseInt(localStorage.userId));
   };
 
   const listOnlineUsers = async (
     params: Users.ListOnlineParams = { limit: 100, offset: 0 },
-    force: boolean = false
+    force: boolean = false,
   ): Promise<Users.User[]> => {
     const { limit, offset, requested_at } = onlineUsersRef.current;
     const currentTimestamp = Date.now();
     const shouldRequest = currentTimestamp - requested_at > 60000;
-    const isDifferentParams =
-      params.limit !== limit || params.offset !== offset;
+    const isDifferentParams = params.limit !== limit || params.offset !== offset;
 
     if (shouldRequest || isDifferentParams || force) {
       try {
@@ -635,9 +546,7 @@ export const ChatProvider = ({
       throw "No dialog provided. You need to provide a dialog via function argument or select a dialog via 'selectDialog'.";
     }
 
-    ConnectyCube.chat.sendIsTypingStatus(
-      dialog.type === 3 ? (getDialogOpponentId(dialog) as number) : dialog._id
-    );
+    ConnectyCube.chat.sendIsTypingStatus(dialog.type === 3 ? (getDialogOpponentId(dialog) as number) : dialog._id);
   };
 
   const getLastActivity = async (userId: number): Promise<string> => {
@@ -661,9 +570,7 @@ export const ChatProvider = ({
           const day = lastLoggedInTime.getUTCDate();
           const month = lastLoggedInTime.getMonth() + 1;
           const year = lastLoggedInTime.getFullYear();
-          status = `Last seen ${day}/${month
-            .toString()
-            .padStart(2, "0")}/${year}`;
+          status = `Last seen ${day}/${month.toString().padStart(2, "0")}/${year}`;
         } else {
           status = `Last seen ${hoursAgo} hours ago`;
         }
@@ -693,12 +600,10 @@ export const ChatProvider = ({
 
   const lastMessageSentTimeString = (dialog: Dialogs.Dialog): string => {
     return formatDistanceToNow(
-      dialog.last_message_date_sent
-        ? (dialog.last_message_date_sent as number) * 1000
-        : (dialog.created_at as string),
+      dialog.last_message_date_sent ? (dialog.last_message_date_sent as number) * 1000 : (dialog.created_at as string),
       {
         addSuffix: true,
-      }
+      },
     );
   };
   const messageSentTimeString = (message: Messages.Message): string => {
@@ -723,7 +628,7 @@ export const ChatProvider = ({
       },
       {
         signal: abortController1.signal,
-      }
+      },
     );
     window.addEventListener(
       "offline",
@@ -733,7 +638,7 @@ export const ChatProvider = ({
       },
       {
         signal: abortController2.signal,
-      }
+      },
     );
 
     return () => {
@@ -750,10 +655,7 @@ export const ChatProvider = ({
 
     // ConnectyCube.chat.onReconnectListener = () => {};
 
-    ConnectyCube.chat.onMessageListener = (
-      userId: number,
-      message: Chat.Message
-    ) => {
+    ConnectyCube.chat.onMessageListener = (userId: number, message: Chat.Message) => {
       // TODO: handle multi-device
       if (userId === currentUserIdRef.current) {
         return;
@@ -762,35 +664,27 @@ export const ChatProvider = ({
       const dialogId = message.dialog_id as string;
       const messageId = message.id;
       const body = message.body || "";
-      const opponentId =
-        message.type === "chat"
-          ? (currentUserIdRef.current as number)
-          : undefined;
+      const opponentId = message.type === "chat" ? (currentUserIdRef.current as number) : undefined;
 
       _stopTyping(userId, dialogId);
 
+      const attachments =
+        message.extension.attachments?.length > 0
+          ? message.extension.attachments.map((attachment: Messages.Attachment) => ({
+              ...attachment,
+              url: ConnectyCube.storage.privateUrl(attachment.uid),
+            }))
+          : undefined;
+
       // add message to store
-      let fileUrl;
-      if (message.extension.attachments?.length > 0) {
-        const fileUID = message.extension.attachments[0].uid;
-        fileUrl = ConnectyCube.storage.privateUrl(fileUID);
-      }
-      _addMessageToStore(
-        messageId,
-        body,
-        dialogId,
-        userId,
-        opponentId,
-        fileUrl
-      );
+      _addMessageToStore(messageId, body, dialogId, userId, opponentId, attachments);
 
       // updates chats store
       setDialogs((prevDialogs) => {
         const dialog = _findDialog(dialogId);
 
         if (!selectedDialog || selectedDialog._id !== message.dialog_id) {
-          dialog.unread_messages_count =
-            (dialog.unread_messages_count || 0) + 1;
+          dialog.unread_messages_count = (dialog.unread_messages_count || 0) + 1;
         }
         dialog.last_message = message.body;
         dialog.last_message_date_sent = parseInt(message.extension.date_sent);
@@ -803,9 +697,7 @@ export const ChatProvider = ({
       }
     };
 
-    ConnectyCube.chat.onSystemMessageListener = async (
-      message: Chat.SystemMessage
-    ) => {
+    ConnectyCube.chat.onSystemMessageListener = async (message: Chat.SystemMessage) => {
       const dialogId = message.extension.dialogId;
       const senderId = message.userId;
 
@@ -833,9 +725,7 @@ export const ChatProvider = ({
         }
         // when someone added new participants to the chat
         case GroupChatEventType.ADD_PARTICIPANTS: {
-          const usersIds = message.extension.addedParticipantsIds
-            .split(",")
-            .map(Number);
+          const usersIds = message.extension.addedParticipantsIds.split(",").map(Number);
           _retrieveAndStoreUsers(usersIds);
           const dialog = _findDialog(dialogId);
 
@@ -847,9 +737,7 @@ export const ChatProvider = ({
         }
         // when someone removed participants from chat
         case GroupChatEventType.REMOVE_PARTICIPANTS: {
-          const usersIds = message.extension.removedParticipantsIds
-            .split(",")
-            .map(Number);
+          const usersIds = message.extension.removedParticipantsIds.split(",").map(Number);
           const dialog = _findDialog(dialogId);
 
           setDialogs((prevDialogs) => {
@@ -876,11 +764,7 @@ export const ChatProvider = ({
       }
     };
 
-    ConnectyCube.chat.onReadStatusListener = (
-      messageId: string,
-      dialogId: string,
-      userId: number
-    ) => {
+    ConnectyCube.chat.onReadStatusListener = (messageId: string, dialogId: string, userId: number) => {
       // TODO: handle multi-device
       if (userId === currentUserIdRef.current) {
         return;
@@ -897,11 +781,7 @@ export const ChatProvider = ({
       });
     };
 
-    ConnectyCube.chat.onMessageTypingListener = (
-      isTyping: boolean,
-      userId: number,
-      dialogId: string
-    ) => {
+    ConnectyCube.chat.onMessageTypingListener = (isTyping: boolean, userId: number, dialogId: string) => {
       // TODO: handle multi-device
       if (userId === currentUserIdRef.current) {
         return;